+++ FROM SCRATCH AGAIN/Assets/FallingObjects.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: b1979f1ea2a9dde4c9c931161fe78edd
-NativeFormatImporter:
+guid: 66533aae06eba42c291172c295972d6d
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 13400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
