+++ FROM SCRATCH AGAIN/Assets/FallingObjects/Bean.prefab
@@ -45,7 +45,7 @@ GameObject:
   - component: {fileID: 3019871721413779049}
   m_Layer: 0
   m_Name: Bean
-  m_TagString: Untagged
+  m_TagString: FallingObject
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
