+++ FROM SCRATCH AGAIN/Assets/FallingObjects/Bean.prefab.meta
