+++ FROM SCRATCH AGAIN/Assets/FallingObjects/FallingObject.cs
@@ -1,0 +1,44 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FallingObject : MonoBehaviour
+{
+    [SerializeField] protected float fallSpeed;
+
+    private void Update()
+    {
+        MoveDownwards();
+    }
+
+    protected virtual void MoveDownwards()
+    {
+        float yTranslation = -fallSpeed * 0.1f * Time.deltaTime;
+        transform.Translate(new Vector3(0, yTranslation, 0));
+    }
+
+    protected void RemoveTerrain(Collider2D collision)
+    {
+        if (collision.gameObject.CompareTag("Terrain"))
+        {
+            collision.gameObject.SetActive(false);
+            gameObject.SetActive(false);
+        }
+    }
+
+    protected void DestroyOnPlayerTouch(Collider2D collision)
+    {
+        if (collision.gameObject.CompareTag("Player"))
+        {
+            gameObject.SetActive(false);
+
+            // Fire an event for adding score, etc
+
+        }
+    }
+
+    protected void DetermineScore()
+    {
+
+    }
+}
