+++ FROM SCRATCH AGAIN/Assets/FallingObjects/FallingObject.prefab
@@ -12,7 +12,7 @@ GameObject:
   - component: {fileID: 2403859125131568722}
   m_Layer: 0
   m_Name: FallingObject
-  m_TagString: Untagged
+  m_TagString: FallingObject
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
