+++ FROM SCRATCH AGAIN/Assets/FallingObjects/HumanRespawner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 48659b76b367e499baa2e4957017f654
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
