+++ FROM SCRATCH AGAIN/Assets/FallingObjects/ObjectPooler.cs
@@ -1,0 +1,62 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ObjectPooler : MonoBehaviour
+{
+    [System.Serializable]
+    public class Pool
+    {
+        public string tag;
+        public GameObject prefab;
+        public int size;
+    }
+
+    public static ObjectPooler Instance;
+
+    private void Awake()
+    {
+        Instance = this;
+    }
+
+    public List<Pool> pools;
+
+    public Dictionary<string, Queue<GameObject>> poolDictionary;
+
+    private void Start()
+    {
+        poolDictionary = new Dictionary<string, Queue<GameObject>>();
+
+        foreach (Pool pool in pools)
+        {
+            Queue<GameObject> objectPool = new Queue<GameObject>();
+
+            for (int i = 0; i < pool.size; i++)
+            {
+                GameObject obj = Instantiate(pool.prefab);
+                obj.SetActive(false);
+                objectPool.Enqueue(obj);
+            }
+
+            poolDictionary.Add(pool.tag, objectPool);
+        }
+    }
+
+    public GameObject SpawnFromPool(string tag, Vector3 spawnPosition)
+    {
+        if (!poolDictionary.ContainsKey(tag))
+        {
+            Debug.LogWarning("Pool with tag " + tag + " does not exist.");
+            return null;
+        }
+
+        GameObject objToSpawn = poolDictionary[tag].Dequeue();
+
+        objToSpawn.SetActive(true);
+        objToSpawn.transform.position = spawnPosition;
+
+        poolDictionary[tag].Enqueue(objToSpawn);
+
+        return objToSpawn;
+    }
+}
