+++ FROM SCRATCH AGAIN/Assets/FallingObjects/ObjectPooler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c4fc13c5c9f448fba4df7914b405e06
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
