--- FROM SCRATCH AGAIN/Assets/KillPlayer.cs
@@ -1,16 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class KillPlayer : MonoBehaviour
-{
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        if (collision.gameObject.CompareTag("Player"))
-        {
-            // Fire an event for game over, etc
-
-            Destroy(collision.gameObject);
-        }
-    }
-}
