+++ FROM SCRATCH AGAIN/Assets/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4d52adcfa46354d44828dbcadfd41a4c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
