+++ OldSpawnManager.cs
